+++ src/components/MenuSider/MenuSider.tsx
@@ -18,6 +18,8 @@ import { checkPermissionRenderMenu } from "../../utility/func/RedirectByPermissi
 import { useRecoilValue } from "recoil";
 import { roleAtom } from "../../store/RoleAtom";
 import packageJson from "../../../package.json";
+import { mockRoles } from "../../utility/StaticPermission";
+import { isArray } from "lodash";
 
 interface Props {
   style?: React.CSSProperties;
@@ -25,18 +27,12 @@ interface Props {
     path: string;
     name: string;
     title: string;
-    permission: {
-      name: string | string[];
-      action: string;
-    } | null;
+    permission: string[];
     subMenu: {
       path: string;
       name: string;
       title: string;
-      permission?: {
-        name: string;
-        action: string;
-      } | null;
+      permission: string;
     }[];
   }[];
   isOpenSidebar?: boolean;
@@ -193,10 +189,6 @@ function MenuSider({ style, lists = [], isOpenSidebar = false }: Props): JSX.Ele
         subPath: isHaveSubPath ? isHaveSubPath.name : "",
       });
     }
-    // const pathNameSplit = pathName.split("/");
-    // if (pathNameSplit.length > 2) {
-    //   setCurrent(pathNameSplit[2]);
-    // }
   });
 
   const onClickList = (name: string) => {
@@ -205,13 +197,20 @@ function MenuSider({ style, lists = [], isOpenSidebar = false }: Props): JSX.Ele
       subPath: "",
     });
   };
+  const newRoles = mockRoles.map((el) => {
+    return {
+      ...el,
+      permission: isArray(el.menu) ? [] : Object.keys(el.menu),
+    };
+  });
+
   return (
     <MenuSiderStyled style={style}>
       <div>
         {lists.map((list, idx) => {
           const isPremiss = checkPermissionRenderMenu({
-            menus: roleData?.menus,
-            permission: list?.permission || null,
+            menus: newRoles,
+            permission: list?.permission || [],
           });
 
           if (!isPremiss) return null;