+++ src/entities/RoleEntity.ts
@@ -2,9 +2,6 @@ export interface RoleCreatePayload {
   rolename?: string;
   roledescription?: string;
   company?: string;
-  menus?: {
-    menuName: string;
-    permission: string[];
-  }[];
+  menus?: any[];
   updateBy?: string;
 }